refactor: extract engine error helpers
